many small fixes for gcc support.
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@3857 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Graphics/GRMesh.cpp
@@ -7,7 +7,7 @@
  */
 #include "Graphics.h"
 #include "GRMesh.h"
-#include <gl/glut.h>
+#include <GL/glut.h>
 
 
 namespace Spr{;
@@ -217,7 +217,7 @@ void GRMesh::DrawBuffer(void* vtx){
 			unsigned from=0, to=0;
 			for(; to<originalFaceIds.size(); ++to){
 				if (originalFaceIds[to] >= materialList.size()) continue;
-				if (materialList[originalFaceIds[to]] == i) continue;
+				if (materialList[originalFaceIds[to]] == (int)i) continue;
 				if (from < to){
 					render->DrawIndexed(GRRenderIf::TRIANGLES, 
 						&*faces.begin() + 3*from, vtx, (to-from)*3);
